+++ controllers/userController.js
@@ -5,6 +5,12 @@ const jwt = require("jsonwebtoken");
 // get all users
 async function getAllUsers(req, res){
     try {
+        const currentUser = req.user; 
+
+        if (currentUser.role !== 'Admin') {
+            return res.status(403).json({ message: 'Forbidden' });
+        }
+
         const users = await User.find({})
 
         if(!users || users.length === 0){
@@ -19,6 +25,13 @@ async function getAllUsers(req, res){
 // get user by id
 async function getUser(req, res){
     try {
+        const currentUser = req.user; 
+
+        if (currentUser.role !== 'Admin') {
+            return res.status(403).json({ message: 'Forbidden' });
+        }
+
+        console.log("user", currentUser)
         const id = req.params.id;
 
         if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -30,6 +43,7 @@ async function getUser(req, res){
         if(!user){
             return res.status(404).json({message: 'User not found'})
         }
+
         return res.status(200).json(user)
     } catch (error) {
         res.status(500).json({error: error.message})
@@ -39,10 +53,12 @@ async function getUser(req, res){
 // create user
 async function createUser(req, res){
     try {
-        const {name, password} = req.body;
+        const {name, password, email, role} = req.body;
         const user = new User({
             name, 
-            password
+            password,
+            email,
+            role
         });
 
         await user.save();
@@ -54,9 +70,9 @@ async function createUser(req, res){
 
 // login
 async function login(req, res) {
-    const { name, password } = req.body;
+    const { email, password } = req.body;
     try {
-        const user = await User.findOne({ name: name });
+        const user = await User.findOne({ email: email });
 
         if (!user) {
             return res.status(401).json({ message: "User not exist" });
@@ -64,8 +80,7 @@ async function login(req, res) {
         if (user.password !== password) {
             return res.status(401).json({ message: "Password incorrect" });
         }
-        // const token = generateToken(email);
-        const token = jwt.sign({ name: name }, "Secret123", { expiresIn: "1h" });
+        const token = jwt.sign({userId: user._id, email: email }, "Secret123", { expiresIn: "1h" });
 
         res.status(200).json(token);
     } catch (error) {
@@ -73,13 +88,18 @@ async function login(req, res) {
     }
 }
 
-
 // update user
 async function updateUser(req, res){
     try {
         const id = req.params.id;
         const filter = {_id: id};
         const update = req.body;
+        const currentUser = req.user;
+
+        if (currentUser.role !== 'Admin' && currentUser._id.toString() !== id) {
+            return res.status(403).json({ message: 'Forbidden' });
+        }
+
         const user = await User.findOneAndUpdate(filter, update)
 
         if(!user){
@@ -96,6 +116,13 @@ async function updateUser(req, res){
 async function deleteUser(req, res){
     try {
         const id = req.params.id;
+
+        const currentUser = req.user;
+
+        if (currentUser.role !== 'Admin' && currentUser._id.toString() !== id) {
+            return res.status(403).json({ message: 'Forbidden' });
+        }
+
         const user = await User.findOneAndDelete({_id: id})
 
         if(!user){
